Update dev branch from master

+++ src/lib/Terminal.svelte
@@ -7,13 +7,13 @@
 	import { spawn, type IPty } from "tauri-pty";
 
 	let terminalElement: HTMLElement;
+	export let hidden = false;
 
 	function initializeXterm() {
 		initShell(terminalElement);
 	}
 
 	onMount(async () => {
-		console.log("mount");
 		initializeXterm();
 	});
 </script>
@@ -31,7 +31,7 @@ export function clearTerminal() {
 }
 
 export function closeTerminal() {
-	pty.kill();
+	if (pty) pty.kill();
 }
 
 export function updateTermTheme() {
@@ -75,11 +75,14 @@ function initShell(terminalElement: HTMLElement) {
 
 <svelte:window on:resize={fitTerminal}></svelte:window>
 
-<div id="terminal" bind:this={terminalElement} />
+<div id="terminal" bind:this={terminalElement} class:hidden />
 
 <style>
 	#terminal {
 		height: 97.5%;
 		width: 100%;
+	}
+	.hidden {
+		display: none;
 	}
 </style>
